+++ src/mmms/spells/AxtalWitherAoe.java
@@ -20,15 +20,15 @@ import net.md_5.bungee.api.ChatColor;
 public class AxtalWitherAoe
 {
 	private Plugin plugin;
-	
+
 	public AxtalWitherAoe(mmbf.main.Main plugin2)
-	{	
+	{
 		plugin = plugin2;
 	}
 
 	int w = -80;
 	Random rand = new Random();
-	
+
 	public boolean onSpell(CommandSender sender, String[] arg)
 	{
 		if (arg.length != 3)
@@ -51,15 +51,15 @@ public class AxtalWitherAoe
 		}
 		if (error)
 			return (true);
-		
+
 		spell(sender, radius, power);
 		return true;
 	}
-	
+
 	public void spell(CommandSender sender, int radius, int power)
 	{
 		Entity launcher = null;
-		
+
 		if (sender instanceof Entity)
 			launcher = (Entity)sender;
 		else if (sender instanceof ProxiedCommandSender)
@@ -88,7 +88,7 @@ public class AxtalWitherAoe
 		animation(radius, lLoc, launcher);
 		deal_damage(radius, power, plist_targets, launcher);
     }
-	
+
 	void		deal_damage(int radius, int power, Player plist[], Entity launcher)
 	{
 		BukkitScheduler scheduler = Bukkit.getServer().getScheduler();
@@ -97,58 +97,58 @@ public class AxtalWitherAoe
             @Override
             public void run()
             {
-            		for (int i = 0; i < 20; i++)
-            		{
-            			if (plist[i] != null)
-            			{
-            				double distance = plist[i].getLocation().distance(launcher.getLocation());
-            				if (distance < radius)
-            				{
-            					int pot_pow = (int)((double)power * ((double)((double)radius - distance) / (double)radius));
-            					plist[i].addPotionEffect((new PotionEffect(PotionEffectType.HARM, 1, pot_pow)));
-            					plist[i].addPotionEffect((new PotionEffect(PotionEffectType.BLINDNESS, 30, 1)));
-            				}
-            			}
-            		}
+					for (int i = 0; i < 20; i++)
+					{
+						if (plist[i] != null)
+						{
+							double distance = plist[i].getLocation().distance(launcher.getLocation());
+							if (distance < radius)
+							{
+								int pot_pow = (int)((double)power * (((double)radius - distance) / (double)radius));
+								plist[i].addPotionEffect((new PotionEffect(PotionEffectType.HARM, 1, pot_pow)));
+								plist[i].addPotionEffect((new PotionEffect(PotionEffectType.BLINDNESS, 30, 1)));
+							}
+						}
+					}
             }
 		};
 		scheduler.scheduleSyncDelayedTask(this.plugin, dealer , 80L);
 	}
-	
+
 	void animation(int radius, Location loc, Entity launcher)
 	{
 		BukkitScheduler scheduler = Bukkit.getServer().getScheduler();
 		Runnable anim_loop = new Runnable() {
             @Override
             public void run() {
-            		Location lloc = launcher.getLocation();
-            		int  n = rand.nextInt(50) + 100;
-            		double precision = n;
-            		double increment = (2 * Math.PI) / precision;
-            		Location particleLoc = new Location(lloc.getWorld(), 0, lloc.getY() + 1.5, 0);
-            		double rad = radius * (w < 0 ? ((double)w / 80) : ((double)w / 5));
-            		double angle = 0;
-            		for(int j = 0; j < precision; j++)
-            		{
-            			angle = (double)j * increment;
-            			particleLoc.setX(lloc.getX() + (rad * Math.cos(angle)));
-            			particleLoc.setZ(lloc.getZ() + (rad * Math.sin(angle)));
-            			particleLoc.setY(lloc.getY() + 1.5);
-            			particleLoc.getWorld().spawnParticle(Particle.SMOKE_LARGE, particleLoc, 1, 0.02, 1.5 * rad, 0.02, 0);
-            		}
-            		if (w < -20 && w % 2 == 0)
-            			particleLoc.getWorld().playSound(particleLoc, Sound.ENTITY_CAT_HISS, (float)((float)radius / 7), (float)(0.5 + ((float)(w + 60) / 100)));
-            		else if (w == -1)
-            		{
-            			particleLoc.getWorld().playSound(particleLoc, Sound.ENTITY_WITHER_SHOOT, (float)((float)radius / 7), 0.77F);
-            			particleLoc.getWorld().playSound(particleLoc, Sound.ENTITY_WITHER_SHOOT, (float)((float)radius / 7), 0.5F);
-            			particleLoc.getWorld().playSound(particleLoc, Sound.ENTITY_WITHER_SHOOT, (float)((float)radius / 7), 0.65F);
-            		}
-            		w++;
+					Location lloc = launcher.getLocation();
+					int  n = rand.nextInt(50) + 100;
+					double precision = n;
+					double increment = (2 * Math.PI) / precision;
+					Location particleLoc = new Location(lloc.getWorld(), 0, lloc.getY() + 1.5, 0);
+					double rad = radius * (w < 0 ? ((double)w / 80) : ((double)w / 5));
+					double angle = 0;
+					for(int j = 0; j < precision; j++)
+					{
+						angle = (double)j * increment;
+						particleLoc.setX(lloc.getX() + (rad * Math.cos(angle)));
+						particleLoc.setZ(lloc.getZ() + (rad * Math.sin(angle)));
+						particleLoc.setY(lloc.getY() + 1.5);
+						particleLoc.getWorld().spawnParticle(Particle.SMOKE_LARGE, particleLoc, 1, 0.02, 1.5 * rad, 0.02, 0);
+					}
+					if (w < -20 && w % 2 == 0)
+						particleLoc.getWorld().playSound(particleLoc, Sound.ENTITY_CAT_HISS, ((float)radius / 7), (float)(0.5 + ((float)(w + 60) / 100)));
+					else if (w == -1)
+					{
+						particleLoc.getWorld().playSound(particleLoc, Sound.ENTITY_WITHER_SHOOT, ((float)radius / 7), 0.77F);
+						particleLoc.getWorld().playSound(particleLoc, Sound.ENTITY_WITHER_SHOOT, ((float)radius / 7), 0.5F);
+						particleLoc.getWorld().playSound(particleLoc, Sound.ENTITY_WITHER_SHOOT, ((float)radius / 7), 0.65F);
+					}
+					w++;
             }
         };
-        
+
         for (int i = -80; i < 5; i++)
-        		scheduler.scheduleSyncDelayedTask(this.plugin, anim_loop , 1L * (i + 81));
+				scheduler.scheduleSyncDelayedTask(this.plugin, anim_loop , 1L * (i + 81));
 	}
 }
